refactor(notifications): move infrastructure files into Infrastructure/ subfolder
Move NotificationsDbContext, NotificationService, Notifier, NotificationsLog,
Channels/*, EntityConfigurations/*, and Jobs/* into Infrastructure/ to start
the feature-folder migration. Update internal using statements; no public
contract changes. See spec: docs/superpowers/specs/2026-05-14-vertical-slice-feature-folders-design.md
