+++ modules/Notifications/src/SimpleModule.Notifications/Infrastructure/Channels/INotificationChannelRegistry.cs
@@ -1,6 +1,6 @@
 using SimpleModule.Notifications.Contracts;
 
-namespace SimpleModule.Notifications.Channels;
+namespace SimpleModule.Notifications.Infrastructure.Channels;
 
 public interface INotificationChannelRegistry
 {
