+++ tests/SimpleModule.Tests.Shared/Fixtures/SimpleModuleWebApplicationFactory.cs
@@ -12,6 +12,7 @@ using SimpleModule.FeatureFlags;
 using SimpleModule.FileStorage;
 using SimpleModule.Host;
 using SimpleModule.Notifications;
+using SimpleModule.Notifications.Infrastructure;
 using SimpleModule.OpenIddict;
 using SimpleModule.OpenIddict.Contracts;
 using SimpleModule.Permissions;
